test(doctor): isolate startup host-link repair
